Minor examples fixes (#546)
* Fix drop method in repositories
* Fix table reference in foreign key
* Update todolist-http-finch README
* Place APIs under http package
* Move TodoListApp to main package
* Sort some imports
* Remove unused implicit
* Support env vars for configuration in finch example
* Update todolist-lib doc
* Update todolist-http-http4s doc
* Update todolist-http-finch doc
* Update examples links in main README.md
* Remove unnecessary import
