+++ README.md
@@ -163,7 +163,10 @@ libraryDependencies += "io.frees" %%% "frees-monix"             % "0.8.0"
 
 ## Freestyle Examples
 
-* [TodoList](./modules/examples/todolist/README.md).
+* [todolist-lib](./modules/examples/todolist-lib/README.md).
+* [todolist-http-http4s](./modules/examples/todolist-http-http4s/README.md).
+* [todolist-http-finch](./modules/examples/todolist-http-finch/README.md).
+* [slick-example](.modules/examples/slick-example/README.md).
 
 ## Commercial Support
