+++ modules/examples/todolist-http-finch/src/main/resources/application.conf
@@ -1,4 +1,6 @@
 http {
   host = "127.0.0.1"
+  host = ${?HOST}
   port = 8081
+  port = ${?PORT}
 }
